Big update
App.js restructure : App.js only handles all the routes to the websites now.

New folders pages and styles:
Everything that used to be in app was moved to homepage.js. Pages will hold onto all the pages that will be displayed and styles holds onto all the css for those pages.

Visual updates:
There is now a wordbox on the right. Everything is currently colored different colors to in order to differentiate where all the boxes are.

Also new npm install required. Downloaded react-router-dom.